DRL_python/requirements.txt
flake8
autoflake
pep8
autopep8

numpy
pandas
matplotlib
tqdm
opencv-python
tensorflow-gpu==1.14
keras==2.1.6
